test: move ws ClientOptions tests to their own file
ws.test.ts has ~30 pre-existing 1s-timeout flakes under debug+ASAN that
are unrelated to this change; keeping the new maxPayload/handshakeTimeout
tests in a separate file lets them run clean in isolation.
